modify serializable error
