+++ app/src/main/java/com/segnities007/seg/Navigation.kt
@@ -1,0 +1,37 @@
+package com.segnities007.seg
+
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import com.segnities007.seg.ui.screens.hub.home.Home
+import com.segnities007.seg.ui.screens.login.Login
+import com.segnities007.seg.ui.screens.splash.Splash
+import kotlinx.serialization.Serializable
+
+@Serializable
+object  Splash
+@Serializable
+object Login
+@Serializable
+object Hub
+
+@Composable
+fun Navigation(
+    navController: NavHostController = rememberNavController(),
+){
+
+    NavHost(navController = navController, startDestination = Login) {
+        composable<Splash> {
+            Splash()
+        }
+        composable<Login> {
+            Login()
+        }
+        composable<Hub>{
+            Home()
+        }
+    }
+
+}
