--- app/src/main/java/com/segnities007/seg/data/model/Navigation.kt
@@ -1,15 +1,0 @@
-package com.segnities007.seg.data.model
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-object  Splash
-
-@Serializable
-object Login
-
-@Serializable
-object Hub
-
-
-
